refactor: repository 패키지를 domain 패키지 안으로 이동
