+++ src/main/java/com/server/crews/applicant/application/ApplicationService.java
@@ -7,9 +7,9 @@ import com.server.crews.applicant.dto.request.ApplicationSaveRequest;
 import com.server.crews.applicant.dto.request.EvaluationRequest;
 import com.server.crews.applicant.dto.response.ApplicationDetailsResponse;
 import com.server.crews.applicant.dto.response.ApplicationsResponse;
-import com.server.crews.applicant.repository.ApplicationRepository;
-import com.server.crews.applicant.repository.NarrativeAnswerRepository;
-import com.server.crews.applicant.repository.SelectiveAnswerRepository;
+import com.server.crews.applicant.domain.repository.ApplicationRepository;
+import com.server.crews.applicant.domain.repository.NarrativeAnswerRepository;
+import com.server.crews.applicant.domain.repository.SelectiveAnswerRepository;
 import com.server.crews.applicant.util.ApplicationMapper;
 import com.server.crews.global.exception.CrewsErrorCode;
 import com.server.crews.global.exception.CrewsException;
@@ -17,9 +17,9 @@ import com.server.crews.global.exception.NotFoundException;
 import com.server.crews.recruitment.domain.NarrativeQuestion;
 import com.server.crews.recruitment.domain.Recruitment;
 import com.server.crews.recruitment.domain.SelectiveQuestion;
-import com.server.crews.recruitment.repository.NarrativeQuestionRepository;
-import com.server.crews.recruitment.repository.RecruitmentRepository;
-import com.server.crews.recruitment.repository.SelectiveQuestionRepository;
+import com.server.crews.recruitment.domain.repository.NarrativeQuestionRepository;
+import com.server.crews.recruitment.domain.repository.RecruitmentRepository;
+import com.server.crews.recruitment.domain.repository.SelectiveQuestionRepository;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Optional;
