+++ src/main/java/com/server/crews/auth/application/AuthService.java
@@ -7,8 +7,8 @@ import com.server.crews.auth.dto.LoginUser;
 import com.server.crews.auth.dto.request.AdminLoginRequest;
 import com.server.crews.auth.dto.request.ApplicantLoginRequest;
 import com.server.crews.auth.dto.response.TokenResponse;
-import com.server.crews.auth.repository.AdministratorRepository;
-import com.server.crews.auth.repository.ApplicantRepository;
+import com.server.crews.auth.domain.repository.AdministratorRepository;
+import com.server.crews.auth.domain.repository.ApplicantRepository;
 import com.server.crews.global.exception.CrewsErrorCode;
 import com.server.crews.global.exception.CrewsException;
 import lombok.RequiredArgsConstructor;
