+++ src/main/java/com/server/crews/auth/application/RefreshTokenService.java
@@ -5,9 +5,9 @@ import com.server.crews.auth.domain.Applicant;
 import com.server.crews.auth.domain.RefreshToken;
 import com.server.crews.auth.domain.Role;
 import com.server.crews.auth.dto.response.TokenResponse;
-import com.server.crews.auth.repository.AdministratorRepository;
-import com.server.crews.auth.repository.ApplicantRepository;
-import com.server.crews.auth.repository.RefreshTokenRepository;
+import com.server.crews.auth.domain.repository.AdministratorRepository;
+import com.server.crews.auth.domain.repository.ApplicantRepository;
+import com.server.crews.auth.domain.repository.RefreshTokenRepository;
 import com.server.crews.global.exception.CrewsException;
 import com.server.crews.global.exception.CrewsErrorCode;
 import lombok.RequiredArgsConstructor;
