+++ src/main/java/com/server/crews/auth/domain/repository/AdministratorRepository.java
@@ -1,4 +1,4 @@
-package com.server.crews.auth.repository;
+package com.server.crews.auth.domain.repository;
 
 import com.server.crews.auth.domain.Administrator;
 import java.util.Optional;
