+++ src/main/java/com/server/crews/global/config/DatabaseInitializer.java
@@ -3,17 +3,17 @@ package com.server.crews.global.config;
 import com.server.crews.applicant.domain.Application;
 import com.server.crews.applicant.domain.NarrativeAnswer;
 import com.server.crews.applicant.domain.SelectiveAnswer;
-import com.server.crews.applicant.repository.ApplicationRepository;
+import com.server.crews.applicant.domain.repository.ApplicationRepository;
 import com.server.crews.auth.domain.Administrator;
 import com.server.crews.auth.domain.Applicant;
-import com.server.crews.auth.repository.AdministratorRepository;
-import com.server.crews.auth.repository.ApplicantRepository;
+import com.server.crews.auth.domain.repository.AdministratorRepository;
+import com.server.crews.auth.domain.repository.ApplicantRepository;
 import com.server.crews.recruitment.domain.Choice;
 import com.server.crews.recruitment.domain.NarrativeQuestion;
 import com.server.crews.recruitment.domain.Recruitment;
 import com.server.crews.recruitment.domain.Section;
 import com.server.crews.recruitment.domain.SelectiveQuestion;
-import com.server.crews.recruitment.repository.RecruitmentRepository;
+import com.server.crews.recruitment.domain.repository.RecruitmentRepository;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
