+++ src/main/java/com/server/crews/recruitment/application/RecruitmentService.java
@@ -5,9 +5,9 @@ import static java.util.stream.Collectors.joining;
 
 import com.server.crews.applicant.domain.Application;
 import com.server.crews.applicant.event.OutcomeDeterminedEvent;
-import com.server.crews.applicant.repository.ApplicationRepository;
+import com.server.crews.applicant.domain.repository.ApplicationRepository;
 import com.server.crews.auth.domain.Administrator;
-import com.server.crews.auth.repository.AdministratorRepository;
+import com.server.crews.auth.domain.repository.AdministratorRepository;
 import com.server.crews.global.CustomLogger;
 import com.server.crews.global.exception.CrewsErrorCode;
 import com.server.crews.global.exception.CrewsException;
@@ -22,9 +22,9 @@ import com.server.crews.recruitment.dto.request.RecruitmentSaveRequest;
 import com.server.crews.recruitment.dto.response.RecruitmentDetailsResponse;
 import com.server.crews.recruitment.dto.response.RecruitmentProgressResponse;
 import com.server.crews.recruitment.dto.response.RecruitmentStateInProgressResponse;
-import com.server.crews.recruitment.repository.NarrativeQuestionRepository;
-import com.server.crews.recruitment.repository.RecruitmentRepository;
-import com.server.crews.recruitment.repository.SelectiveQuestionRepository;
+import com.server.crews.recruitment.domain.repository.NarrativeQuestionRepository;
+import com.server.crews.recruitment.domain.repository.RecruitmentRepository;
+import com.server.crews.recruitment.domain.repository.SelectiveQuestionRepository;
 import com.server.crews.recruitment.util.RecruitmentMapper;
 import java.time.Clock;
 import java.time.LocalDateTime;
