+++ src/test/java/com/server/crews/applicant/application/ApplicationServiceTest.java
@@ -22,9 +22,9 @@ import com.server.crews.applicant.dto.request.EvaluationRequest;
 import com.server.crews.applicant.dto.response.AnswerResponse;
 import com.server.crews.applicant.dto.response.ApplicationDetailsResponse;
 import com.server.crews.applicant.dto.response.SectionAnswerResponse;
-import com.server.crews.applicant.repository.ApplicationRepository;
-import com.server.crews.applicant.repository.NarrativeAnswerRepository;
-import com.server.crews.applicant.repository.SelectiveAnswerRepository;
+import com.server.crews.applicant.domain.repository.ApplicationRepository;
+import com.server.crews.applicant.domain.repository.NarrativeAnswerRepository;
+import com.server.crews.applicant.domain.repository.SelectiveAnswerRepository;
 import com.server.crews.auth.domain.Administrator;
 import com.server.crews.auth.domain.Applicant;
 import com.server.crews.environ.service.ServiceTest;
