+++ src/test/java/com/server/crews/auth/application/AuthServiceTest.java
@@ -11,8 +11,8 @@ import com.server.crews.auth.dto.LoginUser;
 import com.server.crews.auth.dto.request.AdminLoginRequest;
 import com.server.crews.auth.dto.request.ApplicantLoginRequest;
 import com.server.crews.auth.dto.response.TokenResponse;
-import com.server.crews.auth.repository.AdministratorRepository;
-import com.server.crews.auth.repository.ApplicantRepository;
+import com.server.crews.auth.domain.repository.AdministratorRepository;
+import com.server.crews.auth.domain.repository.ApplicantRepository;
 import com.server.crews.environ.service.ServiceTest;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
