+++ src/test/java/com/server/crews/recruitment/domain/repository/RecruitmentRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.server.crews.recruitment.repository;
+package com.server.crews.recruitment.domain.repository;
 
 import static com.server.crews.fixture.RecruitmentFixture.DEFAULT_CODE;
 import static com.server.crews.fixture.RecruitmentFixture.DEFAULT_DEADLINE;
@@ -11,6 +11,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import com.server.crews.auth.domain.Administrator;
 import com.server.crews.environ.repository.RepositoryTest;
 import com.server.crews.recruitment.domain.Recruitment;
+import com.server.crews.recruitment.domain.repository.RecruitmentRepository;
 import jakarta.validation.ConstraintViolationException;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
