+++ src/test/java/com/server/crews/recruitment/domain/repository/SelectiveQuestionRepositoryTest.java
@@ -1,4 +1,4 @@
-package com.server.crews.recruitment.repository;
+package com.server.crews.recruitment.domain.repository;
 
 import static com.server.crews.fixture.QuestionFixture.CHOICES;
 import static com.server.crews.fixture.QuestionFixture.STRENGTH_QUESTION;
@@ -11,6 +11,7 @@ import com.server.crews.environ.repository.RepositoryTest;
 import com.server.crews.recruitment.domain.Recruitment;
 import com.server.crews.recruitment.domain.Section;
 import com.server.crews.recruitment.domain.SelectiveQuestion;
+import com.server.crews.recruitment.domain.repository.SelectiveQuestionRepository;
 import jakarta.validation.ConstraintViolationException;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
